requirements.txt
djangorestframework
django-filter
requests
